+++ src/html/js/players.js
@@ -1,0 +1,37 @@
+var playerUrlPrefix = "https://bitmovin-a.akamaihd.net/bitmovin-player/";
+var playerUrlSuffix = "/bitmovinplayer.js";
+
+var players = {
+  selector: {
+    src: "",
+    desc: "Select a player version"
+  },
+  beta7: {
+    src: "beta/7",
+    desc: "Latest Beta 7"
+  },
+  stable7: {
+    src: "stable/7",
+    desc: "Latest Stable 7"
+  },
+  v7_1_3: {
+    src: "stable/7.1.3",
+    desc: "7.1.3"
+  },
+  v7_2_0: {
+    src: "stable/7.2.0",
+    desc: "7.2.0"
+  },
+  v7_2_8: {
+    src: "stable/7.2.8",
+    desc: "7.2.8"
+  },
+  v7_3_0: {
+    src: "stable/7.3.0",
+    desc: "7.3.0"
+  },
+  v7_5_2: {
+    src: "stable/7.5.2",
+    desc: "7.5.2"
+  }
+}