+++ src/html/js/sources.js
@@ -1,0 +1,32 @@
+var sources = {
+  live: {
+    // hls: "https://m1w-fra.rec.vc/live/ngrp:hB7LUrYwxkOo2dwX_all/master.m3u8",
+    // hls: "https://bitdash-a.akamaihd.net/content/sintel/hls/playlist.m3u8",
+    // hls: "https://mi-msl4-amd.akamaized.net/hls/live/622389/MattStream_en/MattStream_en_all/playlist.m3u8",
+    // hls: "http://sample.vodobox.com/planete_interdite/planete_interdite_alternate.m3u8",
+    // hls: "https://wbc-live-cdn-qa.edge-cdn.net/mi-webcast/ngrp:5ad9b56a0892378fab3f0f12_en_all/master.m3u8",
+    hls: "https://livesegmented-i.akamaihd.net/hls/live/240282/nginxTestStream/nginxTestStream/playlist.m3u8",
+    title: "Live Source"
+  },
+  vod: {
+    hls: "https://videocdnvod1-vh.akamaihd.net/i/5ETW2kuygc19zKr8gpCA9r/7zQAQqSABqatWGgVVZXMD5.aaa,256,596,828,1656,2356,3656,.mp4.csmil/master.m3u8?set-cc-attribute=cc&start=7",
+    title: "VOD Source"
+  },
+  dvrEn: {
+    // hls: "https://streaming-qa.edge-cdn.net/mi-webcast/597093274716efe73b7be65fDvr/playlist.m3u8?DVR",
+    hls: "https://streaming-prod.edge-cdn.net/mi-webcast/5a2fd317340925c36a647402_enDvr/playlist.m3u8?DVR",
+    title: "DVR Source (en)"
+  },
+  dvrDe: {
+    hls: "https://streaming-prod.edge-cdn.net/mi-webcast/5a2fd317340925c36a647402_deDvr/playlist.m3u8?DVR",
+    title: "DVR Source (de)"
+  },
+  dvrFr: {
+    hls: "https://streaming-qa.edge-cdn.net/mi-webcast/5a28f6992d83b4ddb242fdaf_frDvr/playlist.m3u8?DVR",
+    title: "DVR Source (fr)"
+  },
+  dvrJp: {
+    hls: "https://streaming-qa.edge-cdn.net/mi-webcast/5a28f6992d83b4ddb242fdaf_jpDvr/playlist.m3u8?DVR",
+    title: "DVR Source (jp)"
+  }
+};