+++ src/ts/uimanager.ts
@@ -46,6 +46,7 @@ import {Label} from './components/label';
 import PlayerEvent = bitmovin.PlayerAPI.PlayerEvent;
 import {AirPlayToggleButton} from './components/airplaytogglebutton';
 import {PictureInPictureToggleButton} from './components/pictureinpicturetogglebutton';
+import {AudioOnlyOverlay} from './components/audioonlyoverlay';
 import {Spacer} from './components/spacer';
 import {UIUtils} from './uiutils';
 import {ArrayUtils} from './arrayutils';
@@ -465,17 +466,43 @@ export class UIManager {
 
 export namespace UIManager.Factory {
 
-  export function buildDefaultUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
-    return UIManager.Factory.buildModernUI(player, config);
+  export function buildAudioVideoUI(player: PlayerAPI, config: UIConfig = {}): UIManager {    // show smallScreen UI only on mobile/handheld devices
+    let smallScreenSwitchWidth = 600;
+
+    return new UIManager(player, [{
+      ui: modernSmallScreenUI(),
+      condition: (context: UIConditionContext) => {
+        return context.isMobile && context.documentWidth < smallScreenSwitchWidth;
+      },
+    }, {
+      ui: modernUI(),
+    }], config);
   }
 
-  export function buildDefaultSmallScreenUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
-    return UIManager.Factory.buildModernSmallScreenUI(player, config);
+  export function buildAudioOnlyUI(player: PlayerAPI, config: UIConfig = {}): UIManager {   // show smallScreen UI only on mobile/handheld devices
+    let smallScreenSwitchWidth = 600;
+
+    return new UIManager(player, [{
+      ui: modernAudioOnlySmallScreenUI(),
+      condition: (context: UIConditionContext) => {
+        return context.isMobile && context.documentWidth < smallScreenSwitchWidth;
+      },
+    }, {
+      ui: modernAudioOnlyUI(),
+    }], config);
   }
 
-  export function buildDefaultCastReceiverUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
-    return UIManager.Factory.buildModernCastReceiverUI(player, config);
-  }
+  // export function buildDefaultUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
+  //   return UIManager.Factory.buildModernUI(player, config);
+  // }
+
+  // export function buildDefaultSmallScreenUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
+  //   return UIManager.Factory.buildModernSmallScreenUI(player, config);
+  // }
+
+  // export function buildDefaultCastReceiverUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
+  //   return UIManager.Factory.buildModernCastReceiverUI(player, config);
+  // }
 
   function modernUI() {
     let subtitleOverlay = new SubtitleOverlay();
@@ -544,46 +571,119 @@ export namespace UIManager.Factory {
         new PlaybackToggleOverlay(),
         new CastStatusOverlay(),
         controlBar,
-        new TitleBar(),
-        new RecommendationOverlay(),
-        new Watermark(),
+        // new TitleBar(),
+        // new RecommendationOverlay(),
+        // new Watermark(),
         new ErrorMessageOverlay(),
       ],
-      cssClasses: ['ui-skin-modern'],
+      cssClasses: ['ui-skin-mi-wbc'],
     });
   }
 
-  function modernAdsUI() {
-    return new UIContainer({
+  function modernAudioOnlyUI() {
+    let subtitleOverlay = new SubtitleOverlay();
+
+    let settingsPanel = new SettingsPanel({
       components: [
-        new BufferingOverlay(),
-        new AdClickOverlay(),
-        new PlaybackToggleOverlay(),
+        new SettingsPanelItem('Speed', new PlaybackSpeedSelectBox()),
+        new SettingsPanelItem('Audio Track', new AudioTrackSelectBox()),
+        new SettingsPanelItem('Audio Quality', new AudioQualitySelectBox()),
+      ],
+      hidden: true,
+    });
+
+    let subtitleSettingsPanel = new SubtitleSettingsPanel({
+      hidden: true,
+      overlay: subtitleOverlay,
+      settingsPanel: settingsPanel,
+    });
+
+    let subtitleSettingsOpenButton = new SubtitleSettingsOpenButton({
+      subtitleSettingsPanel: subtitleSettingsPanel,
+      settingsPanel: settingsPanel,
+    });
+
+    settingsPanel.addComponent(
+      new SettingsPanelItem(
+        new SubtitleSettingsLabel({text: 'Subtitles', opener: subtitleSettingsOpenButton}),
+        new SubtitleSelectBox()
+    ));
+
+    let controlBar = new ControlBar({
+      components: [
+        settingsPanel,
+        subtitleSettingsPanel,
         new Container({
           components: [
-            new AdMessageLabel({ text: 'Ad: {remainingTime} secs' }),
-            new AdSkipButton(),
+            new PlaybackTimeLabel({ timeLabelMode: PlaybackTimeLabelMode.CurrentTime, hideInLivePlayback: true }),
+            new SeekBar({ label: new SeekBarLabel() }),
+            new PlaybackTimeLabel({ timeLabelMode: PlaybackTimeLabelMode.TotalTime, cssClasses: ['text-right'] }),
           ],
-          cssClass: 'ui-ads-status',
+          cssClasses: ['controlbar-top'],
         }),
-        new ControlBar({
+        new Container({
           components: [
-            new Container({
-              components: [
-                new PlaybackToggleButton(),
-                new VolumeToggleButton(),
-                new VolumeSlider(),
-                new Spacer(),
-                new FullscreenToggleButton(),
-              ],
-              cssClasses: ['controlbar-bottom'],
-            }),
+            new PlaybackToggleButton(),
+            new VolumeToggleButton(),
+            new VolumeSlider(),
+            new Spacer(),
+            new AirPlayToggleButton(),
+            new CastToggleButton(),
+            new SettingsToggleButton({ settingsPanel: settingsPanel }),
+            new FullscreenToggleButton(),
           ],
+          cssClasses: ['controlbar-bottom'],
         }),
       ],
-      cssClasses: ['ui-skin-modern', 'ui-skin-ads'],
+    });
+
+    return new UIContainer({
+      components: [
+        new AudioOnlyOverlay(),
+        subtitleOverlay,
+        new BufferingOverlay(),
+        new PlaybackToggleOverlay(),
+        new CastStatusOverlay(),
+        controlBar,
+        // new TitleBar(),
+        // new RecommendationOverlay(),
+        // new Watermark(),
+        new ErrorMessageOverlay(),
+      ],
+      cssClasses: ['ui-skin-mi-wbc'],
     });
   }
+  // function modernAdsUI() {
+  //   return new UIContainer({
+  //     components: [
+  //       new BufferingOverlay(),
+  //       new AdClickOverlay(),
+  //       new PlaybackToggleOverlay(),
+  //       new Container({
+  //         components: [
+  //           new AdMessageLabel({ text: 'Ad: {remainingTime} secs' }),
+  //           new AdSkipButton(),
+  //         ],
+  //         cssClass: 'ui-ads-status',
+  //       }),
+  //       new ControlBar({
+  //         components: [
+  //           new Container({
+  //             components: [
+  //               new PlaybackToggleButton(),
+  //               new VolumeToggleButton(),
+  //               new VolumeSlider(),
+  //               new Spacer(),
+  //               new FullscreenToggleButton(),
+  //             ],
+  //             cssClasses: ['controlbar-bottom'],
+  //           }),
+  //         ],
+  //       }),
+  //     ],
+  //     cssClasses: ['ui-skin-mi-wbc', 'ui-skin-ads'],
+  //   });
+  // }
 
   function modernSmallScreenUI() {
     let subtitleOverlay = new SubtitleOverlay();
@@ -652,46 +752,54 @@ export namespace UIManager.Factory {
         }),
         settingsPanel,
         subtitleSettingsPanel,
-        new RecommendationOverlay(),
-        new Watermark(),
+        // new RecommendationOverlay(),
+        // new Watermark(),
         new ErrorMessageOverlay(),
       ],
-      cssClasses: ['ui-skin-modern', 'ui-skin-smallscreen'],
+      cssClasses: ['ui-skin-mi-wbc', 'ui-skin-smallscreen'],
     });
   }
 
-  function modernSmallScreenAdsUI() {
-    return new UIContainer({
+  function modernAudioOnlySmallScreenUI() {
+    let subtitleOverlay = new SubtitleOverlay();
+
+    let settingsPanel = new SettingsPanel({
       components: [
-        new BufferingOverlay(),
-        new AdClickOverlay(),
-        new PlaybackToggleOverlay(),
-        new TitleBar({
-          components: [
-            // dummy label with no content to move buttons to the right
-            new Label({ cssClass: 'label-metadata-title' }),
-            new FullscreenToggleButton(),
-          ],
-        }),
-        new Container({
-          components: [
-            new AdMessageLabel({ text: 'Ad: {remainingTime} secs' }),
-            new AdSkipButton(),
-          ],
-          cssClass: 'ui-ads-status',
-        }),
+        new SettingsPanelItem('Speed', new PlaybackSpeedSelectBox()),
+        new SettingsPanelItem('Audio Track', new AudioTrackSelectBox()),
+        new SettingsPanelItem('Audio Quality', new AudioQualitySelectBox()),
       ],
-      cssClasses: ['ui-skin-modern', 'ui-skin-ads', 'ui-skin-smallscreen'],
+      hidden: true,
+      hideDelay: -1,
     });
-  }
 
-  function modernCastReceiverUI() {
+    let subtitleSettingsPanel = new SubtitleSettingsPanel({
+      hidden: true,
+      hideDelay: -1,
+      overlay: subtitleOverlay,
+      settingsPanel: settingsPanel,
+    });
+
+    let subtitleSettingsOpenButton = new SubtitleSettingsOpenButton({
+      subtitleSettingsPanel: subtitleSettingsPanel,
+      settingsPanel: settingsPanel,
+    });
+
+    settingsPanel.addComponent(
+      new SettingsPanelItem(
+        new SubtitleSettingsLabel({text: 'Subtitles', opener: subtitleSettingsOpenButton}),
+        new SubtitleSelectBox()
+    ));
+
+    settingsPanel.addComponent(new CloseButton({ target: settingsPanel }));
+    subtitleSettingsPanel.addComponent(new CloseButton({ target: subtitleSettingsPanel }));
+
     let controlBar = new ControlBar({
       components: [
         new Container({
           components: [
             new PlaybackTimeLabel({ timeLabelMode: PlaybackTimeLabelMode.CurrentTime, hideInLivePlayback: true }),
-            new SeekBar({ smoothPlaybackPositionUpdateIntervalMs: -1 }),
+            new SeekBar({ label: new SeekBarLabel() }),
             new PlaybackTimeLabel({ timeLabelMode: PlaybackTimeLabelMode.TotalTime, cssClasses: ['text-right'] }),
           ],
           cssClasses: ['controlbar-top'],
@@ -699,198 +807,265 @@ export namespace UIManager.Factory {
       ],
     });
 
-    return new CastUIContainer({
-      components: [
-        new SubtitleOverlay(),
-        new BufferingOverlay(),
-        new PlaybackToggleOverlay(),
-        new Watermark(),
-        controlBar,
-        new TitleBar({ keepHiddenWithoutMetadata: true }),
-        new ErrorMessageOverlay(),
-      ],
-      cssClasses: ['ui-skin-modern', 'ui-skin-cast-receiver'],
-    });
-  }
-
-  export function buildModernUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
-    // show smallScreen UI only on mobile/handheld devices
-    let smallScreenSwitchWidth = 600;
-
-    return new UIManager(player, [{
-      ui: modernSmallScreenAdsUI(),
-      condition: (context: UIConditionContext) => {
-        return context.isMobile && context.documentWidth < smallScreenSwitchWidth && context.isAdWithUI;
-      },
-    }, {
-      ui: modernAdsUI(),
-      condition: (context: UIConditionContext) => {
-        return context.isAdWithUI;
-      },
-    }, {
-      ui: modernSmallScreenUI(),
-      condition: (context: UIConditionContext) => {
-        return context.isMobile && context.documentWidth < smallScreenSwitchWidth;
-      },
-    }, {
-      ui: modernUI(),
-    }], config);
-  }
-
-  export function buildModernSmallScreenUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
-    return new UIManager(player, [{
-      ui: modernSmallScreenAdsUI(),
-      condition: (context: UIConditionContext) => {
-        return context.isAdWithUI;
-      },
-    }, {
-      ui: modernSmallScreenUI(),
-    }], config);
-  }
-
-  export function buildModernCastReceiverUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
-    return new UIManager(player, modernCastReceiverUI(), config);
-  }
-
-  function legacyUI() {
-    let settingsPanel = new SettingsPanel({
-      components: [
-        new SettingsPanelItem('Video Quality', new VideoQualitySelectBox()),
-        new SettingsPanelItem('Audio Track', new AudioTrackSelectBox()),
-        new SettingsPanelItem('Audio Quality', new AudioQualitySelectBox()),
-        new SettingsPanelItem('Subtitles', new SubtitleSelectBox()),
-      ],
-      hidden: true,
-    });
-
-    let controlBar = new ControlBar({
-      components: [
-        settingsPanel,
-        new PlaybackToggleButton(),
-        new SeekBar({ label: new SeekBarLabel() }),
-        new PlaybackTimeLabel(),
-        new VRToggleButton(),
-        new VolumeControlButton(),
-        new SettingsToggleButton({ settingsPanel: settingsPanel }),
-        new CastToggleButton(),
-        new FullscreenToggleButton(),
-      ],
-    });
-
     return new UIContainer({
       components: [
-        new SubtitleOverlay(),
+        new AudioOnlyOverlay(),
+        subtitleOverlay,
+        new BufferingOverlay(),
         new CastStatusOverlay(),
         new PlaybackToggleOverlay(),
-        new Watermark(),
-        new RecommendationOverlay(),
         controlBar,
-        new TitleBar(),
-        new ErrorMessageOverlay(),
-      ],
-      cssClasses: ['ui-skin-legacy'],
-    });
-  }
-
-  function legacyAdsUI() {
-    return new UIContainer({
-      components: [
-        new AdClickOverlay(),
-        new ControlBar({
+        new TitleBar({
           components: [
-            new PlaybackToggleButton(),
-            new AdMessageLabel(),
-            new VolumeControlButton(),
+            new MetadataLabel({ content: MetadataLabelContent.Title }),
+            new CastToggleButton(),
+            new VRToggleButton(),
+            new VolumeToggleButton(),
+            new SettingsToggleButton({ settingsPanel: settingsPanel }),
             new FullscreenToggleButton(),
           ],
         }),
-        new AdSkipButton(),
-      ],
-      cssClasses: ['ui-skin-legacy', 'ui-skin-ads'],
-    });
-  }
-
-  function legacyCastReceiverUI() {
-    let controlBar = new ControlBar({
-      components: [
-        new SeekBar(),
-        new PlaybackTimeLabel(),
-      ],
-    });
-
-    return new UIContainer({
-      components: [
-        new SubtitleOverlay(),
-        new PlaybackToggleOverlay(),
-        new Watermark(),
-        controlBar,
-        new TitleBar(),
+        settingsPanel,
+        subtitleSettingsPanel,
+        // new RecommendationOverlay(),
+        // new Watermark(),
         new ErrorMessageOverlay(),
       ],
-      cssClasses: ['ui-skin-legacy', 'ui-skin-cast-receiver'],
+      cssClasses: ['ui-skin-mi-wbc', 'ui-skin-smallscreen'],
     });
   }
 
-  function legacyTestUI() {
-    let settingsPanel = new SettingsPanel({
-      components: [
-        new SettingsPanelItem('Video Quality', new VideoQualitySelectBox()),
-        new SettingsPanelItem('Audio Track', new AudioTrackSelectBox()),
-        new SettingsPanelItem('Audio Quality', new AudioQualitySelectBox()),
-        new SettingsPanelItem('Subtitles', new SubtitleSelectBox()),
-      ],
-      hidden: true,
-    });
+  // function modernSmallScreenAdsUI() {
+  //   return new UIContainer({
+  //     components: [
+  //       new BufferingOverlay(),
+  //       new AdClickOverlay(),
+  //       new PlaybackToggleOverlay(),
+  //       new TitleBar({
+  //         components: [
+  //           // dummy label with no content to move buttons to the right
+  //           new Label({ cssClass: 'label-metadata-title' }),
+  //           new FullscreenToggleButton(),
+  //         ],
+  //       }),
+  //       new Container({
+  //         components: [
+  //           new AdMessageLabel({ text: 'Ad: {remainingTime} secs' }),
+  //           new AdSkipButton(),
+  //         ],
+  //         cssClass: 'ui-ads-status',
+  //       }),
+  //     ],
+  //     cssClasses: ['ui-skin-mi-wbc', 'ui-skin-ads', 'ui-skin-smallscreen'],
+  //   });
+  // }
 
-    let controlBar = new ControlBar({
-      components: [settingsPanel,
-        new PlaybackToggleButton(),
-        new SeekBar({ label: new SeekBarLabel() }),
-        new PlaybackTimeLabel(),
-        new VRToggleButton(),
-        new VolumeToggleButton(),
-        new VolumeSlider(),
-        new VolumeControlButton(),
-        new VolumeControlButton({ vertical: false }),
-        new SettingsToggleButton({ settingsPanel: settingsPanel }),
-        new CastToggleButton(),
-        new FullscreenToggleButton(),
-      ],
-    });
+  // function modernCastReceiverUI() {
+  //   let controlBar = new ControlBar({
+  //     components: [
+  //       new Container({
+  //         components: [
+  //           new PlaybackTimeLabel({ timeLabelMode: PlaybackTimeLabelMode.CurrentTime, hideInLivePlayback: true }),
+  //           new SeekBar({ smoothPlaybackPositionUpdateIntervalMs: -1 }),
+  //           new PlaybackTimeLabel({ timeLabelMode: PlaybackTimeLabelMode.TotalTime, cssClasses: ['text-right'] }),
+  //         ],
+  //         cssClasses: ['controlbar-top'],
+  //       }),
+  //     ],
+  //   });
 
-    return new UIContainer({
-      components: [
-        new SubtitleOverlay(),
-        new CastStatusOverlay(),
-        new PlaybackToggleOverlay(),
-        new Watermark(),
-        new RecommendationOverlay(),
-        controlBar,
-        new TitleBar(),
-        new ErrorMessageOverlay(),
-      ],
-      cssClasses: ['ui-skin-legacy'],
-    });
-  }
+  //   return new CastUIContainer({
+  //     components: [
+  //       new SubtitleOverlay(),
+  //       new BufferingOverlay(),
+  //       new PlaybackToggleOverlay(),
+  //       new Watermark(),
+  //       controlBar,
+  //       new TitleBar({ keepHiddenWithoutMetadata: true }),
+  //       new ErrorMessageOverlay(),
+  //     ],
+  //     cssClasses: ['ui-skin-mi-wbc', 'ui-skin-cast-receiver'],
+  //   });
+  // }
 
-  export function buildLegacyUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
-    return new UIManager(player, [{
-      ui: legacyAdsUI(),
-      condition: (context: UIConditionContext) => {
-        return context.isAdWithUI;
-      },
-    }, {
-      ui: legacyUI(),
-    }], config);
-  }
+  // export function buildModernUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
+  //   // show smallScreen UI only on mobile/handheld devices
+  //   let smallScreenSwitchWidth = 600;
 
-  export function buildLegacyCastReceiverUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
-    return new UIManager(player, legacyCastReceiverUI(), config);
-  }
+  //   return new UIManager(player, [{
+  //     ui: modernSmallScreenAdsUI(),
+  //     condition: (context: UIConditionContext) => {
+  //       return context.isMobile && context.documentWidth < smallScreenSwitchWidth && context.isAdWithUI;
+  //     },
+  //   }, {
+  //     ui: modernAdsUI(),
+  //     condition: (context: UIConditionContext) => {
+  //       return context.isAdWithUI;
+  //     },
+  //   }, {
+  //     ui: modernSmallScreenUI(),
+  //     condition: (context: UIConditionContext) => {
+  //       return context.isMobile && context.documentWidth < smallScreenSwitchWidth;
+  //     },
+  //   }, {
+  //     ui: modernUI(),
+  //   }], config);
+  // }
 
-  export function buildLegacyTestUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
-    return new UIManager(player, legacyTestUI(), config);
-  }
+  // export function buildModernSmallScreenUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
+  //   return new UIManager(player, [{
+  //     ui: modernSmallScreenAdsUI(),
+  //     condition: (context: UIConditionContext) => {
+  //       return context.isAdWithUI;
+  //     },
+  //   }, {
+  //     ui: modernSmallScreenUI(),
+  //   }], config);
+  // }
+
+  // export function buildModernCastReceiverUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
+  //   return new UIManager(player, modernCastReceiverUI(), config);
+  // }
+
+  // function legacyUI() {
+  //   let settingsPanel = new SettingsPanel({
+  //     components: [
+  //       new SettingsPanelItem('Video Quality', new VideoQualitySelectBox()),
+  //       new SettingsPanelItem('Audio Track', new AudioTrackSelectBox()),
+  //       new SettingsPanelItem('Audio Quality', new AudioQualitySelectBox()),
+  //       new SettingsPanelItem('Subtitles', new SubtitleSelectBox()),
+  //     ],
+  //     hidden: true,
+  //   });
+
+  //   let controlBar = new ControlBar({
+  //     components: [
+  //       settingsPanel,
+  //       new PlaybackToggleButton(),
+  //       new SeekBar({ label: new SeekBarLabel() }),
+  //       new PlaybackTimeLabel(),
+  //       new VRToggleButton(),
+  //       new VolumeControlButton(),
+  //       new SettingsToggleButton({ settingsPanel: settingsPanel }),
+  //       new CastToggleButton(),
+  //       new FullscreenToggleButton(),
+  //     ],
+  //   });
+
+  //   return new UIContainer({
+  //     components: [
+  //       new SubtitleOverlay(),
+  //       new CastStatusOverlay(),
+  //       new PlaybackToggleOverlay(),
+  //       new Watermark(),
+  //       new RecommendationOverlay(),
+  //       controlBar,
+  //       new TitleBar(),
+  //       new ErrorMessageOverlay(),
+  //     ],
+  //     cssClasses: ['ui-skin-legacy'],
+  //   });
+  // }
+
+  // function legacyAdsUI() {
+  //   return new UIContainer({
+  //     components: [
+  //       new AdClickOverlay(),
+  //       new ControlBar({
+  //         components: [
+  //           new PlaybackToggleButton(),
+  //           new AdMessageLabel(),
+  //           new VolumeControlButton(),
+  //           new FullscreenToggleButton(),
+  //         ],
+  //       }),
+  //       new AdSkipButton(),
+  //     ],
+  //     cssClasses: ['ui-skin-legacy', 'ui-skin-ads'],
+  //   });
+  // }
+
+  // function legacyCastReceiverUI() {
+  //   let controlBar = new ControlBar({
+  //     components: [
+  //       new SeekBar(),
+  //       new PlaybackTimeLabel(),
+  //     ],
+  //   });
+
+  //   return new UIContainer({
+  //     components: [
+  //       new SubtitleOverlay(),
+  //       new PlaybackToggleOverlay(),
+  //       new Watermark(),
+  //       controlBar,
+  //       new TitleBar(),
+  //       new ErrorMessageOverlay(),
+  //     ],
+  //     cssClasses: ['ui-skin-legacy', 'ui-skin-cast-receiver'],
+  //   });
+  // }
+
+  // function legacyTestUI() {
+  //   let settingsPanel = new SettingsPanel({
+  //     components: [
+  //       new SettingsPanelItem('Video Quality', new VideoQualitySelectBox()),
+  //       new SettingsPanelItem('Audio Track', new AudioTrackSelectBox()),
+  //       new SettingsPanelItem('Audio Quality', new AudioQualitySelectBox()),
+  //       new SettingsPanelItem('Subtitles', new SubtitleSelectBox()),
+  //     ],
+  //     hidden: true,
+  //   });
+
+  //   let controlBar = new ControlBar({
+  //     components: [settingsPanel,
+  //       new PlaybackToggleButton(),
+  //       new SeekBar({ label: new SeekBarLabel() }),
+  //       new PlaybackTimeLabel(),
+  //       new VRToggleButton(),
+  //       new VolumeToggleButton(),
+  //       new VolumeSlider(),
+  //       new VolumeControlButton(),
+  //       new VolumeControlButton({ vertical: false }),
+  //       new SettingsToggleButton({ settingsPanel: settingsPanel }),
+  //       new CastToggleButton(),
+  //       new FullscreenToggleButton(),
+  //     ],
+  //   });
+
+  //   return new UIContainer({
+  //     components: [
+  //       new SubtitleOverlay(),
+  //       new CastStatusOverlay(),
+  //       new PlaybackToggleOverlay(),
+  //       new Watermark(),
+  //       new RecommendationOverlay(),
+  //       controlBar,
+  //       new TitleBar(),
+  //       new ErrorMessageOverlay(),
+  //     ],
+  //     cssClasses: ['ui-skin-legacy'],
+  //   });
+  // }
+
+  // export function buildLegacyUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
+  //   return new UIManager(player, [{
+  //     ui: legacyAdsUI(),
+  //     condition: (context: UIConditionContext) => {
+  //       return context.isAdWithUI;
+  //     },
+  //   }, {
+  //     ui: legacyUI(),
+  //   }], config);
+  // }
+
+  // export function buildLegacyCastReceiverUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
+  //   return new UIManager(player, legacyCastReceiverUI(), config);
+  // }
+
+  // export function buildLegacyTestUI(player: PlayerAPI, config: UIConfig = {}): UIManager {
+  //   return new UIManager(player, legacyTestUI(), config);
+  // }
 }
 
 export interface SeekPreviewArgs extends NoArgs {